+++ vowpalwabbit/cache.h
@@ -16,7 +16,7 @@ char* run_len_encode(char *p, size_t i);
 
 int read_cached_features(void*a, example* ec);
 void cache_tag(io_buf& cache, v_array<char> tag);
-void cache_features(io_buf& cache, example* ae, uint32_t mask);
+void cache_features(io_buf& cache, example* ae, uint64_t mask);
 void output_byte(io_buf& cache, unsigned char s);
-void output_features(io_buf& cache, unsigned char index, feature* begin, feature* end, uint32_t mask);
+void output_features(io_buf& cache, unsigned char index, feature* begin, feature* end, uint64_t mask);
