+++ vowpalwabbit/ftrl.cc
@@ -43,7 +43,7 @@ void multipredict(ftrl& b, base_learner&, example& ec, size_t count, size_t step
   for (size_t c=0; c<count; c++)
     pred[c].scalar = ec.l.simple.initial;
   GD::multipredict_info mp = { count, step, pred, &all.reg, (float)all.sd->gravity };
-  GD::foreach_feature<GD::multipredict_info, uint32_t, GD::vec_add_multipredict>(all, ec, mp);
+  GD::foreach_feature<GD::multipredict_info, uint64_t, GD::vec_add_multipredict>(all, ec, mp);
   if (all.sd->contraction != 1.)
     for (size_t c=0; c<count; c++)
       pred[c].scalar *= (float)all.sd->contraction;
