+++ vowpalwabbit/global_data.h
@@ -114,7 +114,7 @@ typedef float weight;
 
 struct regressor
 { weight* weight_vector;
-  size_t weight_mask; // (stride*(1 << num_bits) -1)
+  uint64_t weight_mask; // (stride*(1 << num_bits) -1)
   uint32_t stride_shift;
 };
 
@@ -456,7 +456,7 @@ struct vw
   size_t pass_length;
   size_t numpasses;
   size_t passes_complete;
-  size_t parse_mask; // 1 << num_bits -1
+  uint64_t parse_mask; // 1 << num_bits -1
   bool permutations; // if true - permutations of features generated instead of simple combinations. false by default
   v_array<v_string> interactions; // interactions of namespaces to cross.
   std::vector<std::string> pairs; // pairs of features to cross.
@@ -473,7 +473,7 @@ struct vw
   uint32_t skips[256];//skips in ngrams.
   std::vector<std::string> limit_strings; // descriptor of feature limits
   uint32_t limit[256];//count to limit features by
-  uint32_t affix_features[256]; // affixes to generate (up to 8 per namespace)
+  uint64_t affix_features[256]; // affixes to generate (up to 8 per namespace)
   bool     spelling_features[256]; // generate spelling features for which namespace
   vector<string> dictionary_path;  // where to look for dictionaries
   vector<feature_dict*> namespace_dictionaries[256]; // each namespace has a list of dictionaries attached to it
